Proyecto01/requirements.txt
numpy
matplotlib
pyserial
pywin32